Relocate and extend role-stripping
